move examples into subfolders
