+++ examples/xlnstorch/full_examples/mnist_example.py
